server/comms,dex/ws: increase websocket read limits
This increases the server's authorized read limit by a factor of 4.
We have observed the previous limit of 65 KiB is inadequate when there
are a large number of orders in a single epoch, and with more funding
coins than normal.

This also doubles the unauthorized/default read limit, which is still
quite small at 8192 bytes.