client/{core,asset/btc}: add RescanWallet and implement for btc spv
Add (*spvWallet).rescanWalletAsync.

Create ExchangeWalletSPV, which embeds ExchangeWallet, and adds the
Rescan method that makes it into an asset.Rescanner.

Refactor out SyncStatus loop into a method, walletCheckAndNotify, that
is to be run as a goroutine from both connectWallet and RescanWallet.
For clear goroutine semantics at the call site, the caller must update
Core's WaitGroup and define the closure that calls walletCheckAndNotify
and wg.Done afterwards.