client/asset/{btc,dcr}: reset redeemScripts when changing min confs
In the UTXO selection methods for DCR and BTC, initially only UTXOs
with >0 confs are tried, and if there are insufficient funds, it
attempts it again with unconfirmed utxos included (minus untrusted
for BTC). However, on the second run, both methods neglect to reset
the redeemScripts slices.  This fixes the issue.