client: order options
Add order-time wallet options. The options are returned as part of
a pre-order estimate, and are persisted as part of order metadata
so that they can be passed to FundOrder, Swap, and Redeem. BTC,
clones, and DCR all get two swap options (fee bump, split) and
1 redemption option (fee bump).

Define two different option types, boolean and xy-range.
Each input type corresponds to a different UI input widget, with
boolean being a simple checkbox while xy-range offers a slider and
text inputs. The types are intentionally generic for reusability.