dex,client/asset: add token primitives
dex:
Token and Gases are used to specify token info in higher-level
packages, where lgpl builds are not suitable.
Add a test token ID.

client/asset:
The Token type embeds dex.Token and adds a WalletDefinition. This
type is used to register tokens with the package-level RegisterToken
method.

client/asset/eth:
The test token is registered with it's WalletDefinition. The
definition has 1 ConfigOption, limitAllowance, that allows the user
to keep a limited allowance for the swap contract, but with extra
transactions = more fees.