client/asset/dcr: coin not found not an error for spv
Do not return asset.CoinNotFoundError from SwapConfirmations with an
SPV wallet since this is normal for mempool transactions in SPV.

For a non-SPV wallet, if the swap appears spent but it cannot be
located in a block with a cfilters scan, this will return
asset.CoinNotFoundError. For SPV wallets, it is not an error if the
transaction cannot be located SPV wallets cannot see non-wallet
transactions until they are mined.

Fix a few logs where arguments were orderd strangely.

Also silence a warning that happens each time a transaction is
found in a block while looking for a spending transaction immediately
afterward.