eth: Fix swap and redeem.
On the client side:

Allow decoding coins to discern between tx hashes and funding coin
encodings.

Temporarily set max fees to 200 so that they can be mined on our simnet.

Return tx hashes for redeems instead of inputs because that's currently
what the server needs to create swap coins.

Add a new error type for uninitiated swaps. This is returned when
checking swap confirmations and locktime for logging purposes so this is
not considered an error for logging purposes.

Add the ability to squelch eth's internal node because it is very loud
on trace.

For the server:

Return TxData and SecretHash with contract data. TxData is the entire
serialized transaction that initiates a swap.