client/asset,core: SwapConfirmations tx not found handling
client/asset/btc: Also return CoinNotFoundError for SPV wallets when
they cannot locate the transaction.  If the wallet is not SPV, log a
warning when SwapConfirmations has this result.

client/asset/dcr: If the wallet SPV log a warning when SwapConfirmations
when the result is CoinNotFound.

client/core: In counterPartyConfirms, do not log an error from
SwapConfirmations if it is asset.CoinNotFound.

Allow CoinNotFound to log an error when checking own swap.