eth swap contract v1
Don't store contract data on-chain, just commit to it in the key of
the swaps map, which maps to a "swap record", 32 bytes in length,
with a special interpretation scheme that allows us to track the
swap status. This reduces fees by a factor of ~2 for a single lot
round-trip trade, which would be similar to e.g. ETH-USDC or other
token trade. For many-lot trades, the savings are even better,
asymptoting at ~28% of the v0 fees. A single-lot round trip trade
uses ~ 93k gas, which is notably less than a trade on UniSwap and
other Ethereum trading platforms.