web/markets: Remove comma from field int value.
Doc's formatCoinValue will stick a comma in the value that we save to an
int field, causing it not to save at all, preventing four or more digit
coin values from being traded with my language settings. Prevent this by
using a simpler conversion that does not insert localized punctuation.