eth: supplement eth swap receipt
This supplements the swapRecipt String method with the contract address,
which in addition to the secret hash makes performing a manual refund
more straightforward.

This also omits the raw refund transaction logging in
(*Core).swapMatchGroup when the Receipt.SignedRefund provides no such
data, as is the case with ETH.
The swap Receipt's Stringer is still used here to log the secret hash
and contract address for ETH swapReceipts.