client/core: allow wallet Lock to timeout
For various reasons, some unknown, it is apparently quite common for 
wallet lock to hang either client shutdown or just logout. This is 
particularly problematic because it leaves the user no choice but to 
hard kill the process, which may have undesirable consequences. Even 
when it is just slow to lock, most users decide it's hung forever much 
sooner. I'm estimating this threshold at around 5-10 seconds.

This change adds a timeout argument to wallet.Lock. The individual 
implementations could also be modified with say a context.Context or a 
time.Duration argument, but in cases where the backend is not context 
enabled it ends up like this anyway, just with the backend worrying 
about this and the caller (core) being ultimately at the mercy of the 
backends actually doing it, back to possible hangs.

Also of note is that this change makes wallet.Lock errors no longer 
prevent full Logout (and thus clean shutdown) possible. This was only 
encouraging dangerous killing of the process.