client/asset/btc: unlock outpoint on sendRawTransaction
When sendRawTransaction returns with a non-nil error, even if we hit the
PublishTransaction timeout (defaultBroadcastWait), we still need to
unlock the outpoint in the wallet.
Although listunspent no longer returns the spent outpoint, the
lockedSats computations reference the locked outpoints listed by
listlockunspent, and they are not removed when spent.