client/asset: Refund must indicate when spent
It is critical that asset.Wallet Refund signal to caller that it is
spent via asset.CoinNotFoundError so that it knows to begin looking
for the counterparty's redeem and move on to redeem as well.

This fixes dcr's refundTx so that it wraps asset.CoinNotFoundError
when lookupTxOutput indicates it is spent.

This also makes explicit the requirement for the asset.Wallet.Refund
method to return asset.CoinNotFoundError when spent.