client/core: allow more Refund errors to trigger redemption search
When locktime is expired, isRefundable will be true and thus block
the shouldBeginFindRedemption case in tick.  We handle this case
by having Refund errors signal when it is spent and then trigger a
redemption search with findMakersRedemption.  This improves
the docs on this handling and the need for Refund to return
asset.CoinNotFoundError when spent. This change also starts
findMakersRedemption for any non-nil error if it is a taker match
and the status is TakerSwapCast, but it does not block refund retries.