client/ui/btc: Configure wallet bday
The wallet birthday for the bitcoin SPV wallet was previously
always set to a hardcoded value. This diff stores the time when
the app seed was generated in the db to give an earliest possible
time so scan from, and also adds options on the UI to specify the
desired wallet birthday. Also, the wallet birthday can now be updated
when reconfiguring the wallet.

* Update spvWallet.birthday when doing rescan

* Automatic rescans and block bday update if there are active orders