client/core: try a fresh fee rate for the init
When creating the swap transaction, see if a more current fee rate
should be used instead of the prescribed swap rate. This can be
particularly helpful for the taker because their swap may be a
long time after the match is made because the maker's swap must
be fully confirmed first.

If the current rate is swapRate < currentRate <= MaxFeeRate, then
use the current rate for the swap transaction.