server/swap: wait for init/redeem for full wait expiration
In handleInit and handleRedeem, do not bother to limit the expiration
time of the processInit/processRedeem waiters by the broadcast timeout.
Both of the TryFuncs (processInit and processRedeem) will respond in
error if the match is revoked, so there is no need to tweak the waiter's
expire time.

This will prevent a client request timeout when the init or redeem is
submitted just before the Swapper revokes the match, when the waiter's
expire time could be set to the past despite the inaction checks having
not yet revoked the match.