server/swap: revoke when contracts are/would be expired
This updates Swapper to proactively revoke matches where any known swap
contracts are expired, or in the case of MakerSwapCast where the
expected taker contract would be created with a lock time in the past.
These scenarios are possible if a transaction takes an extremely long
time to reach the required number of confirmations.

The new checks are done in checkInactionEventBased, which is called
on a ticker.

The (*Swapper).failMatch method is given a userFault bool input
argument to indicate if the actor should be penalized.

matchTracker is given an expiredBy method that returns true if the lock
time of either party's *known* swap is before the reference time.
Normally this will be called with the present time.

checkInactionBlockBased now consults expiredBy when calling failMatch
to determine if either of the swap contracts are expired, justifying
the inaction.

NOTE: client should be updated with similar logic to prevent attempting
to make expired contracts or contracts that are about to expire.

server/db: option to store forgiven revoked matches

When a match is revoked due to no fault of the user, it should be saved
in the DB with the forgiven flag set.