client/asset/btc: attempt audit with empty txdata
When AuditContract was switched to a hard requirement of txData
(the raw tx bytes), it broke certain simnet tests where no server was
available to provide the data, and made the asset.Wallet API somewhat
incomplete with no way to retrieve this txData in either full node or
SPV mode because there is no method to do so.  This makes a redeem
cli tool or other use case with no server or txdata oracle extremely
cumbersome because the raw bytes are required as an input instead of
just the coinid/txid.

Normally the server should communicate this txData, and the caller can
decide to require it. This adds (back) the ability to work with an
empty txData as a convenience for recovery tools and testing.  This MAY
change in the future if a GetTxData method is added for this purpose.

The baseWallet AuditContract now attempts retrieval of the transaction
output from the network, but it is only ensured to succeed for a full
node or, if the tx is confirmed, and SPV wallet.  If this is a full node
wallet, a simple gettxout RPC is sufficient with no pkScript or "since"
time. If this is an SPV wallet, only a confirmed counterparty contract
can be located, and only one within the last 48 hours. As such, this
mode of operation is not intended for normal server-coordinate
operation.

The above change fixes the "livetest" package, which is also used by
all btc "clone" assets. The test code is also updated with some extra
sleeps for SPV wallets.  A comment at the call site of AuditContract
is made noting that a nil txData must be provided because in normal
operation that data comes from the server coordinating the swap, and
that the Wallet API has no ability to get this data.