server/swap: fix memory leak with coin locker
Because transactions can advance swap status before procesBlock has a
chance to tryConfirmSwap and then unlockOrderCoins, there is a memory
leak when we see the "Prematurely received" warnings from processInit
and processRedeem.

This moves the unlockOrderCoins calls from revoke to deleteMatch, where
it is called for all matches, including completed ones. (see the
MatchComplete branch in processRedeem).