client/core: add onion proxy option
Add the Onion config field, for use with .onion DEX hosts.
If Onion is not set, the TorProxy value is used instead.

connectDEX recognizes .onion hosts and uses the Onion proxy.

In connectDEX, the scheme (ws or wss) is determined by
the presence of a TLS certificate.  Only .onion hosts are
permitted to use a no-TLS connection ("ws" scheme).

parseCert is updated so that a nil certI interface{} is permitted, in
which case it checks the cert store, but a nil []byte may be expected
if the host is a .onion service.