client/asset/dcr: fix side chain and approval checks
The mainchain and valid returns from (*ExchangeWallet).blockHeader were
switched.  This flips them correctly.

This also changes the way a block is detected as side chain by checking
the Confirmations field of the getblockverbose result instead of
requesting the hash of the block at the same height and comparing
the hashes.  A side chain block will have -1 for Confirmations.

This also changes the way the next block is retrieved for stakeholder
approval checking.  Instead of requesting the hash of the block at the
next height, use the NextHash field returned by getblockverbose.
This saves an RPC, but it is also more robust since there is no need
to use strings.Contains on the error returned from getblockhash
to detect if there is no next block (we are at tip).  If we are at tip,
the NextHash field is empty.  With getblockhash, the wallet says
"-1: Block number out of range: N" while the node says
"-4: wallet.BlockInfo: item does not exist: no block at height N in main
chain".  Although client only ever checks with wallet, these strings are
fairly arbitrary  human readable messages that are likely to change
without warning.