client: Electrum exchange wallet (BTC, LTC, BCH)
electrum: add wallet client

Add the HTTP JSON-RPC client for making requests to the local
Electrum wallet instance.

electrum: add network/server client

Add a TCP JSON-RPC client for making requests to remote ElectrumX
servers.

client/asset/btc: define externalAddress in asset.Wallet

This replaces the separate addressPKH and addressWPKH interface methods
with a single externalAddress method. This matches the single
changeAddress method.

The asset.Wallet implementations, such as rpcWallet and spvWallet, know
their segwit capability and mode. In the case of rpcWallet, it has a
segwit flag so that it can support multiple bitcoind backends with or
without segwit addresses. In the case of spvWallet, it only ever
provides segwit addresses and does all operations assuming segwit,
and it had addressPKH returning an "unimplemented" error.
The case of Electrum will be similar to rpcWallet, with a segwit flag
to indicate what sort of external wallet is being driven.

client/asset/btc: move the RPC-oriented error helpers to RPC code

This moves the rpcclient-oriented error helpers, isTxNotFoundErr and
isMethodNotFoundErr, into rpcclient.go where they should only be used.
Errors from other wallets such as the native SPV wallet and an Electrum
wallet should not use these functions as they match on the
dcrjson.RPCError type and btcjson error codes, which are only applicable
to a bitcoin core backend and an rpcclient client.

Instead of using isTxNotFoundErr in baseWallet after calling
node.getWalletTransaction to translate to asset.CoinNotFoundError, this
now translates as appropriate for the asset.Wallet implementation.
Namely, in rpcclient.go it uses these existing helper functions, while
in spv.go it checks the WalletTransactionNotFound error kind.

Unfortunately, isMethodNotFoundErr is stil applied in the baseWallet
constructor when calling the estimateFee helper to detect backends
that do not offer the method.  This is not as easily changed because
the estimateFee helper accepts a RawRequester and is inherently
RPC-oriented (but some implementers just ignore that argument).
This is something to refactor another day.

client/asset/btc: add ElectrumExchangeWallet

This creates the ExchangeWalletElectrum type for using an external
Electrum wallet along with an ElectrumX server.

This is intended to support Electrum forks for LTC, BCH, and BTC.

The new walletTypeElectrum is defined as "electrumRPC".

The ExchangeWalletElectrum type overrides some baseWallet methods:
- FindRedemption since the process is very different (and simpler)
  from that used by block filter-enabled wallets.
- feeRate, to avoid pointless conversions and btcjson types
- watchBlocks and reportNewTip, since it is not necessary to retarget a
  starting point for redemption searches.
- Connect, since it sets estimateFee to the overridden feeRate method,
  and starts the overridden watchBlocks instead of the baseWallet
  method.

It also statisfies the Sweeper and FeeRater interfaces.

To power ExchangeWalletElectrum, we define an electrumWallet type
that satisfies the asset.Wallet interface similar to rpcWallet and
spvWallet.

The electrumWallet type is powered by an electrumWalletClient, which
communicates with the external Electrum wallet via RPC, and an
electrumNetworkClient, which communicates with a remote Electrum(X)
server for block headers and verbose transaction data.

The general design of electrumWallet is to govern electrumNetworkClient
so that it tracks the electrum rever in used by the Electrum *wallet*.
This is done for a couple of reasons:
- ensure consistent results between the two
- allow the Electrum wallet to handle server selection

-----

add getBestBlockHeader to asset.Wallet

This introduces to the asset.Wallet interface a getBestBlockHeader
method that can be used in place of getBestBlockHash followed by
blockFromHash.

This method is helpful since not all backends allow querying for
block headers by hash. (I'm looking at you Electrum.)

-----

add a fast checkWalletTx optional wallet backend method

This defines an optional walletTxChecker interface, with a single
checkWalletTx method, to complement the current asset.Wallet interface.
walletTxChecker provide a fast wallet tx query when block info not
needed. This should be treated as an optimization method, where
getWalletTransaction may always be used in its place if it does not
exists.

The difference between getWalletTransaction and checkWalletTx is that
getWalletTransaction will return a more verbose struct containing
details on the block containing the transaction. If it is only
necessary to obtain the raw transaction and a confirmation count,
the checkWalletTx may be used.
This is also motivated by Electrum wallets where it is considerably
more expensive to obtain the verbose result from a network server
than to get the checkWalletTx result from the local wallet DB.

As with getWalletTransaction, checkWalletTx is only intended to work
with wallet related transactions.

-----

remove Details field from GetTransactionResult

This simplifies the GetTransactionResult type in wallettypes.go.
Several fields from the bitcoind/btcwallet response that are both
unneeded and difficult to satisfy are now omitted. e.g. the native
btcwallet and an external Electrum backend would have to jump through
hoops to reproduce certain fields that are not strictly necessary.
Namely, the Amount, Fee, and Details fields. All of the required
data can be obtained by decoding the raw transaction contined in the
Hex field.

tolerate some address reuse where needed

client/cmd/simnet-trade-tests: add wallet type replacing spv bool