client/core/btc/ui: Accelerate Orders using CPFP
Adds support for Accelerating BTC transactions using the 
Child-Pays-For-Parent technique. Adds a new Accelerator wallet trait 
which includes three functions: AccelerateOrder which creates a new 
transaction sending money to the wallet's change address with a high fee 
in order to expedite mining, PreAccelerate, which gives guidelines to 
the user, and AccelerationEstimate which returns how much it will cost 
to accelerate the order to a certain fee rate.

On the UI, the orders page is updated with a button which shows up if 
the order is able to be accelerated. When clicking the button, a popup 
shows up which allows the user to accelerate an order.