client/ui/btc: Recover BTC SPV wallet
This adds the ability to delete all the files from the BTC SPV wallet and
recreate it from scratch.

client/asset/interface: A Recoverer trait is added that includes three
functions:
- GetRecoveryCfg: Attempts to retrieve a map[string]string from the
  existing wallet that would help in putting the wallet back to the
  state is was in before all the wallet files are moved.
- Move: move all the wallet files to a backup folder.
    
client/asset/btc: Implements the Recoverer interface. The recovery 
configuration includes the number of internal and number of external 
addresses that the wallet generated. This amount of addresses are 
regenerated in the new instance of the wallet. If this was not done and 
there were active orders being managed when the user recovered the 
wallet, the new wallet would not be able to redeem and swaps because it 
wouldn't have access to the required private keys.

client/core: A Recover function is added that destroys the old wallet 
and creates a new one. It contains a force parameter similar to Rescan 
which unless is set to true, will not allow the user to recover a wallet 
if there are currently active deals requiring the wallet. An 
activeOrdersErr is returned from the wallet if force=false and there are 
currently active deals.

app: A button is added on the wallet's page to allow users to recover a 
wallet if the wallet has the Recoverer trait. The "force" workflow is 
implemented for both rescan and recover. If either of these requests 
return the activeOrdersErr, another popup will show up prompting the 
user to confirm that they would like to take this action even though 
there are active orders. If the user confirms, the same request will be 
resubmitted with force=true.