update btc harness
On LTC regtest/regnet, the mweb soft fork is not active at genesis.
It is necessary to mine to block 431, send to a mweb address in a
"peg-in" transaction, and then mine the block so that the pegin
and the required hogex may be created when mweb activates at 432.

As of Bitcoin Core 0.21, the default "" wallet is not automatically
created.
https://github.com/bitcoin/bitcoin/blob/master/doc/release-notes/release-notes-0.21.0.md#automatic-wallet-creation-removed
This adds a CREATE_DEFAULT_WALLET variable to explicitly create
this wallet if needed.  The command also creates it encrypted to start.