client/asset/ltc: work with and require v0.21.2
This updates the client's supported Litecoin Core version to v0.21.2

This is the modern Litecoin Core, with the MWEB patch set.

In addition to bumping the version, this means:
- getbalances is supported now
- sendrawtransaction has the new syntax with maxfeerate not allowHighFee

This also adds the missing switches on wallet Type to both ltc and bch.

To support MWEB and MW txns, this defines custom block and transaction
deserializers.  This is based on the LIPS:
https://github.com/litecoin-project/lips/blob/master/lip-0002.mediawiki
https://github.com/litecoin-project/lips/blob/master/lip-0003.mediawiki

On LTC regtest/regnet, the mweb soft fork is not active at genesis.
It is necessary to mine to block 431, send to a mweb address in a
"peg-in" transaction, and then mine the block so that the pegin
and the required hogex may be created when mweb activates at 432.
As such this also regenerates the litecoin harness chain tarball.

As of Bitcoin Core 0.21, the default "" wallet is not automatically
created.
https://github.com/bitcoin/bitcoin/blob/master/doc/release-notes/release-notes-0.21.0.md#automatic-wallet-creation-removed
This adds a CREATE_DEFAULT_WALLET variable to the harness script
to explicitly create this wallet if needed.  The command also creates it
encrypted to start.