client/core,webserver,rpcserver: order submit time
Use order submission time, which is a local time stamp, instead of the
server's time stamp, which is from the server's system clock, when
displaying order age and time on the frontend.

This adds the core.Order.SubmitTime field to complement the Stamp
field.  This marshals to json as "submitTime".

The frontend is updated to use "submitTime" instead of "stamp".