eth: Wait for tx in tests.
When running the simnet tests on testnet, we wait for all tx to be mined
before proceeding. Sometimes, on testnet, this coincides with us looking
at a to be orphaned block. When we try to find the transaction that was
just mined, it may not be mined in our block. As a solution, wait a few
blocks to find tx in tests. If we wait until we are back on the main
chain, we will find it.

While we are here, update the internal geth nodes to use geth v1.10.20.