client/{asset,core}: add UnlockCoinsOnLogin option
This adds the core.Config.UnlockCoinsOnLogin option that indicates that
on wallet connect during login, or on creation of a new wallet, all
coins with the wallet should be unlocked.

To support this "unlock all" with the wallets, the ReturnCoins method
of the asset.Wallet interface now interprets a nil Coins slice to mean
that all coins should be unlocked. This is similar to the lockunspent
RPC's semantics.

The individual implementations of ReturnCoins are modified to recognize
a nil unspent input slice and accordingly request the wallet backend
unlock all coins.