client/eth: Use wallet seed as entropy for BIP-39 mnemonic
Previously, the wallet seed from core was used as the master seed
of the an HD wallet from which a private key was derived. Instead
of this, the seed from core is used as an entropy for a BIP-39
mnemonic phrase, then the seed derived from the mnemonic phrase
is used to derive the private key used in the wallet. The mnemonic
phrase is currenlty not used for anything, but this will keep our
options open for more features in the future.