core: refactor tick and matchTracker access model
This is a bigger change to core to deal with the deadlocks and lock 
contention with `trackedTrades`.  The main changes are:

- Only _read_ locking the `trackedTrade` mutex for the portion of `tick` 
where the checks are going on (with lots of wallet requests), and only 
write locking when actions are taken subsequently that may need to 
modify trade and/or match status and metadata.  This is the section of 
code that was causing delays and outright hangs in many places, namely 
anywhere that was using `(*trackedTrade).isActive`, which is all over 
the place.

- More fine grained concurrency controls for some `matchTracker` fields 
to allow certain mutations while the parent `trackedTrade`'s mutex is 
_not_ write locked, such as during the checking stage of `tick` and 
other places that don't modify any of the actual `trackedTrade` fields.  
The parent's mutex is still the main guard for most of the matches 
fields, just now with some exceptions.

- Preventing concurrent runs of `tick` for the same `trackedTrade` using 
a `tickLock` mutex at the start of `tick`.  This is necessary now 
that the entirety of `tick` is not done with the `trackedTrade`'s mutex 
*write* locked.  The possibility for concurrent ticks arises because 
tick can be called from at least three different spots, in response to 
different events, and thus all on different goroutines.