server/{matcher,market}: base start and end rate calcs on actual match rates
We have always used the mid-gap at the beginning and end of the epoch
for our candle's start rate and end rate. In a traditional market, you would
use the first and last match price. Use the expected numbers instead.
The candle should also not go to zero on an empty book anymore.