client/core: bond messaging and account tier
This adapts Core to handle the updated msgjson types. In particular,
recognizing the Tier and other bonds fields. Legacy and bond-enabled
servers are both supported.

The config response now interprets the presence of bond info to mean
the effective API version is BondAPIVersion. Such servers continue
to advertise v0 (PreAPIVersion) until we decide to purge the unstable
API and force client updgrades.