client/core: unlock coins and reconcile failed epoch status orders
When reconciling and order that was in epoch status when the connection
was lost and not restored before the preimage request, be sure to unlock
funding coins.

On reconnect, queue any epoch status orders for status resolution
after the expected time of matching. They are revoked if they remain
in epoch status. This prevents uncancelable orders that do not resolve
until a subsequent reconnect or restart.

This also improves "stale cancel order" cleanup by checking for and
removing such orders in tick. This is is needed since the existing call
to deleteStaleCancelOrder from reconcileTrades can be called too soon
to classify the cancel order as stale even if it is.