build: Use targets and loop in pkg.sh.
This modifies the packaging script to make use of a TARGETS variable to
control which platforms are built rather than having to manually specify
the OS and ARCH multiple times for each command.

This should be less error prone when modifying the target platforms and
is certainly easier to add/remove platforms as desired.