client/core: purge legacy preimage request handling
The initial preimage request did not include the order commitment for
the preimage being requested.  This had us doing some tricks because
the request only provided the order ID, which is not known to the client
until the order submission response was received, creating a race.

Since this flawed design was resolved in 850e8a6, in v0.2.0 about 18
months ago, by including the commitment in the preimage request,
we can safely remove the legacy client-side handling for requests
that do not include the commitment.  This change removes the piSyncers
map and legacy code path from handlePreimageRequest.