core/eth: Handle block gas limit
Two new fields are added to `WalletInfo`: `MaxSwapsInTx` and
`MaxRedeemsInTx`. `Core` then uses these fields to call `Swap`
and `Redeem` in batches if needed. These fields are currently
only populated by ETH and ETH token wallets, and they check each
of the versions of the contract, and the fields are populated so that
at most 1/4 of the gas limit of a block can be filled up by a swap or
redeem transaction.