client/core: allow earlier auto-redeem as taker
This modifies handleRedemptionRoute so that if the order ID in the
request is not active, the DB is checked for the match. If all of: it
is found, we are taker, and our own redeem coin is recorded, we respond
to the request since it is required as part of the settlement sequence.
As maker, we only record the taker's redeem coin provided in the
request, however this request is not presently sent from server to
maker because settlement is completed.

Now that we can handle 'redemption' requests from the server for orders
that have been retired, we can as taker redeem as soon as we see the
maker's redeem of our swap by starting a redemption search. Previously,
we would wait for a very long time for the redemption request to make
our redeem, which was frustrating behavior that we required since
performing a redemption search and auto-redeem implied both
self-revoking the match as well as retiring the order after the redeem.
This commit makes these factors less important, simply delaying by
a short but reasonable duration after first observing the swap is
spent before performing the redemption search.